requirements.txt
# Add your project dependencies here
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
agents>=0.1.0
